Compatibility of dataio and FitsReadingModule with IFS data, new WavelengthReadingModule, updated ParangReadingModule
