+++ .readthedocs.yml
@@ -1,5 +1,12 @@
+version: 2
+
+sphinx:
+  configuration: docs/conf.py
+
 build:
     image: latest
 
 python:
-    version: 3.6
+   version: 3.7
+   install:
+      - requirements: requirements.txt
