+++ pynpoint/__init__.py
@@ -78,6 +78,13 @@ from pynpoint.processing.timedenoising import CwtWaveletConfiguration, \
                                               WaveletTimeDenoisingModule, \
                                               TimeNormalizationModule
 
+from pynpoint.readwrite.attr_reading import AttributeReadingModule, \
+                                            ParangReadingModule, \
+                                            WavelengthReadingModule
+
+from pynpoint.readwrite.attr_writing import AttributeWritingModule, \
+                                            ParangWritingModule
+
 from pynpoint.readwrite.fitsreading import FitsReadingModule
 
 from pynpoint.readwrite.fitswriting import FitsWritingModule
@@ -86,12 +93,7 @@ from pynpoint.readwrite.hdf5reading import Hdf5ReadingModule
 
 from pynpoint.readwrite.hdf5writing import Hdf5WritingModule
 
-from pynpoint.readwrite.textwriting import AttributeWritingModule, \
-                                           ParangWritingModule, \
-                                           TextWritingModule
-
-from pynpoint.readwrite.textreading import ParangReadingModule, \
-                                           AttributeReadingModule
+from pynpoint.readwrite.textwriting import TextWritingModule
 
 from pynpoint.readwrite.nearreading import NearReadingModule
