+++ pynpoint/core/attributes.py
@@ -89,6 +89,10 @@ def get_attributes():
                        'config': 'header',
                        'value': 'None',
                        'type': 'float'},
+            'WAVELENGTH': {'attribute': 'non-static',
+                           'config': 'header',
+                           'value': 'None',
+                           'type': 'float'},
             'STAR_POSITION': {'attribute': 'non-static',
                               'config': None,
                               'value': None,
