+++ setup.py
@@ -18,7 +18,8 @@ setup(
     long_description_content_type='text/x-rst',
     author='Tomas Stolker & Markus Bonse',
     author_email='tomas.stolker@phys.ethz.ch',
-    url='http://pynpoint.ethz.ch',
+    url='https://github.com/PynPoint/PynPoint',
+    project_urls={'Documentation': 'https://pynpoint.readthedocs.io'},
     packages=['pynpoint',
               'pynpoint.core',
               'pynpoint.readwrite',
