+++ tests/test_readwrite/test_textwriting.py
@@ -6,8 +6,7 @@ import pytest
 import numpy as np
 
 from pynpoint.core.pypeline import Pypeline
-from pynpoint.readwrite.textwriting import TextWritingModule, ParangWritingModule, \
-                                           AttributeWritingModule
+from pynpoint.readwrite.textwriting import TextWritingModule
 from pynpoint.util.tests import create_config, create_random, remove_test_data
 
 warnings.simplefilter('always')
@@ -28,8 +27,7 @@ class TestTextReading:
 
     def teardown_class(self):
 
-        remove_test_data(self.test_dir, files=['image.dat', 'parang.dat',
-                                               'attribute.dat', 'data.dat'])
+        remove_test_data(self.test_dir, files=['image.dat', 'data.dat'])
 
     def test_input_data(self):
 
@@ -95,74 +93,3 @@ class TestTextReading:
 
         assert np.allclose(data, data_int, rtol=limit, atol=0.)
         assert data.shape == (100, )
-
-    def test_parang_writing(self):
-
-        parang_write = ParangWritingModule(file_name='parang.dat',
-                                           name_in='parang_write1',
-                                           output_dir=None,
-                                           data_tag='images',
-                                           header=None)
-
-        self.pipeline.add_module(parang_write)
-        self.pipeline.run_module('parang_write1')
-
-        data = np.loadtxt(self.test_dir+'parang.dat')
-
-        assert np.allclose(data[0], 1.0, rtol=limit, atol=0.)
-        assert np.allclose(data[9], 10.0, rtol=limit, atol=0.)
-        assert np.allclose(np.mean(data), 5.5, rtol=limit, atol=0.)
-        assert data.shape == (10, )
-
-    def test_attribute_writing(self):
-
-        attr_write = AttributeWritingModule(file_name='attribute.dat',
-                                            name_in='attr_write1',
-                                            output_dir=None,
-                                            data_tag='images',
-                                            attribute='PARANG',
-                                            header=None)
-
-        self.pipeline.add_module(attr_write)
-        self.pipeline.run_module('attr_write1')
-
-        data = np.loadtxt(self.test_dir+'attribute.dat')
-
-        assert np.allclose(data[0], 1.0, rtol=limit, atol=0.)
-        assert np.allclose(data[9], 10.0, rtol=limit, atol=0.)
-        assert np.allclose(np.mean(data), 5.5, rtol=limit, atol=0.)
-        assert data.shape == (10, )
-
-    def test_attribute_not_present(self):
-
-        attr_write = AttributeWritingModule(file_name='attribute.dat',
-                                            name_in='attr_write3',
-                                            output_dir=None,
-                                            data_tag='images',
-                                            attribute='test',
-                                            header=None)
-
-        self.pipeline.add_module(attr_write)
-
-        with pytest.raises(ValueError) as error:
-            self.pipeline.run_module('attr_write3')
-
-        assert str(error.value) == 'The \'test\' attribute is not present in \'images\'.'
-
-    def test_parang_writing_not_present(self):
-
-        with h5py.File(self.test_dir+'PynPoint_database.hdf5', 'a') as hdf_file:
-            del hdf_file['header_images/PARANG']
-
-        parang_write = ParangWritingModule(file_name='parang.dat',
-                                           name_in='parang_write3',
-                                           output_dir=None,
-                                           data_tag='images',
-                                           header=None)
-
-        self.pipeline.add_module(parang_write)
-
-        with pytest.raises(ValueError) as error:
-            self.pipeline.run_module('parang_write3')
-
-        assert str(error.value) == 'The PARANG attribute is not present in \'images\'.'
